requirements.txt
gunicorn==20.1.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
opencv-python-headless==4.8.1.78
packaging==23.2
pillow==10.1.0
PyMuPDF==1.23.5
pytesseract==0.3.10
Werkzeug==3.0.1
redis==4.5.4
prometheus-client==0.16.0
python-dotenv==1.0.0